Add a temporary system call to pring a message

+++ kern/proc.h
@@ -48,7 +48,7 @@ enum procstate { UNUSED, EMBRYO, SLEEPING, RUNNABLE, RUNNING, ZOMBIE };
 
 // Per-process state
 struct proc {
-  uint sz;              // Size of process memory (bytes)
+  size_t sz;            // Size of process memory (bytes)
   pte_t *pgdir;         // Page table
   char *kstack;         // Bottom of kernel stack for this process
   enum procstate state; // Process state
